requirements.txt
asgiref==3.11.1
Django==6.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
dj-database-url==1.0.0
whitenoise==6.5.0
requests==2.31.0
python-dateutil==2.8.2
Pillow==10.2.0

